change delete prices method

+++ app/models/suplemento.rb
@@ -27,7 +27,7 @@ class Suplemento < ApplicationRecord
 
   def delete_old_prices(query)
     prices = self.prices.order(:created_at)
-    while prices.count > 30
+    if prices.count > 30
         prices.first(prices.size - 30).delete_all
         puts "old prices deleted for #{self.name}"
     end
